+++ ci/test_all_features.sh
@@ -2,5 +2,5 @@
 set -euxo pipefail
 
 for feature in $(tomljson Cargo.toml | jq --raw-output '.features | keys[]' | grep -v 'default\|testing-helpers'); do
-    cargo test --tests --no-default-features --features "$feature,testing-helpers";
+    cargo test -p derive_more --tests --no-default-features --features "$feature,testing-helpers";
 done
