+++ impl/doc/into_iterator.md
