+++ impl/doc/is_variant.md
