+++ impl/src/is_variant.rs
