+++ impl/src/mul_assign_like.rs
