+++ impl/src/mul_helpers.rs
