+++ impl/src/mul_like.rs
