+++ impl/src/utils.rs
@@ -1,4 +1,8 @@
-#![cfg_attr(not(feature = "default"), allow(dead_code), allow(unused_mut))]
+#![cfg_attr(
+    not(all(feature = "add", feature = "mul")),
+    allow(dead_code),
+    allow(unused_mut)
+)]
 
 use proc_macro2::{Span, TokenStream};
 use quote::{quote, ToTokens};
